fixed typo
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10826 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ oscam-garbage.c
@@ -28,7 +28,7 @@ static int32_t garbage_collector_active;
 static int32_t garbage_debug;
 
 #ifdef WITH_DEBUG
-void add_garbage_debug(void *data, char *file, uint16_t line)
+void add_garbage_debug(void *data, char *file, uint32_t line)
 {
 #else
 void add_garbage(void *data)
